# course-social-network/course-network-ui/src/app/services/models/borrowed-course-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BorrowedCourseResponse {
  authorName?: string;
  id?: number;
  isbn?: string;
  rate?: number;
  achieveApproved?: boolean;
  achieved?: boolean;
  title?: string;
}