# course-social-network/course-network-ui/src/app/services/models/course-response.ts
/* tslint:disable */
/* eslint-disable */
export interface CourseResponse {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  isbn?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}