# course-social-network/course-network-ui/src/app/services/models/page-response-borrowed-course-response.ts
/* tslint:disable */
/* eslint-disable */
import { BorrowedCourseResponse } from './borrowed-course-response';
export interface PageResponseBorrowedCourseResponse {
  content?: Array<BorrowedCourseResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}