# course-social-network/course-network-ui/src/app/services/models/page-response-course-response.ts
/* tslint:disable */
/* eslint-disable */
import { CourseResponse } from './course-response';
export interface PageResponseCourseResponse {
  content?: Array<CourseResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}