# course-social-network/course-network-ui/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification - Ines","description":"OpenApi documentation for Spring Security","termsOfService":"Terms of service","contact":{"name":"Ines","url":"https://inescoding.com/course","email":"contact@inescoding.com"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"Local ENV"},{"url":"https://inescoding.com/course","description":"PROD ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/feedbacks":{"post":{"tags":["Feedback"],"operationId":"saveFeedback","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/courses":{"get":{"tags":["Course"],"operationId":"findAllCourses","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseCourseResponse"}}}}}},"post":{"tags":["Course"],"operationId":"saveCourse","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/courses/cover/{course-id}":{"post":{"tags":["Course"],"operationId":"uploadCourseCoverPicture","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/courses/borrow/{course-id}":{"post":{"tags":["Course"],"operationId":"borrowCourse","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/courses/shareable/{course-id}":{"patch":{"tags":["Course"],"operationId":"updateShareableStatus","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/courses/borrow/return/{course-id}":{"patch":{"tags":["Course"],"operationId":"returnBorrowCourse","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/courses/borrow/return/approve/{course-id}":{"patch":{"tags":["Course"],"operationId":"approveReturnBorrowCourse","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/courses/archived/{course-id}":{"patch":{"tags":["Course"],"operationId":"updateArchivedStatus","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/feedbacks/course/{course-id}":{"get":{"tags":["Feedback"],"operationId":"findAllFeedbacksByCourse","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseFeedbackResponse"}}}}}}},"/courses/{course-id}":{"get":{"tags":["Course"],"operationId":"findCourseById","parameters":[{"name":"course-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseResponse"}}}}}}},"/courses/owner":{"get":{"tags":["Course"],"operationId":"findAllCoursesByOwner","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseCourseResponse"}}}}}}},"/courses/borrowed":{"get":{"tags":["Course"],"operationId":"findAllBorrowedCourses","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedCourseResponse"}}}}}}},"/courses/achieved":{"get":{"tags":["Course"],"operationId":"findAllAchievedCourses","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseBorrowedCourseResponse"}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"FeedbackRequest":{"required":["comment","courseId"],"type":"object","properties":{"note":{"maximum":5,"minimum":0,"type":"number","format":"double"},"comment":{"type":"string"},"courseId":{"type":"integer","format":"int32"}}},"CourseRequest":{"required":["authorName","isbn","synopsis","title"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"synopsis":{"type":"string"},"shareable":{"type":"boolean"}}},"RegistrationRequest":{"required":["email","firstname","lastname","password"],"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"FeedbackResponse":{"type":"object","properties":{"note":{"type":"number","format":"double"},"comment":{"type":"string"},"ownFeedback":{"type":"boolean"}}},"PageResponseFeedbackResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/FeedbackResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"CourseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"synopsis":{"type":"string"},"owner":{"type":"string"},"cover":{"type":"array","items":{"type":"string","format":"byte"}},"rate":{"type":"number","format":"double"},"archived":{"type":"boolean"},"shareable":{"type":"boolean"}}},"PageResponseCourseResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"BorrowedCourseResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"authorName":{"type":"string"},"isbn":{"type":"string"},"rate":{"type":"number","format":"double"},"achieved":{"type":"boolean"},"achieveApproved":{"type":"boolean"}}},"PageResponseBorrowedCourseResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/BorrowedCourseResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}