# course-social-network/course-network/src/main/java/com/ProjectRH/course/course/BorrowedCourseResponse.java
package com.ProjectRH.course.course;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedCourseResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean achieved;
    private boolean achieveApproved;
}